# laravel/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "API documentation for my project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API server"
        }
    ],
    "paths": {
        "/api/comments": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Create a new comment",
                "description": "Store a newly created comment in storage.",
                "operationId": "5e13e1318a258179a078e21595139cc5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CommentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/comments/{id}": {
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Update comment",
                "description": "Update the specified comment in storage.",
                "operationId": "fff220031d937469c766c2593d86555f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CommentResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comment updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Delete comment",
                "description": "Remove the specified comment from storage.",
                "operationId": "2ad57afe280d350b97a89252a605fa07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Comment deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreCommentRequest": {
                "required": [
                    "user_id",
                    "commentable_type",
                    "commentable_id",
                    "text"
                ],
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "commentable_type": {
                        "type": "string",
                        "enum": [
                            "App\\Services\\News\\Models\\News",
                            "App\\Services\\VideoPost\\Models\\VideoPost",
                            "App\\Services\\Comment\\Models\\Comment"
                        ],
                        "example": "App\\Services\\News\\Models\\News"
                    },
                    "commentable_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "text": {
                        "type": "string",
                        "example": "This is a comment"
                    }
                },
                "type": "object"
            },
            "UpdateCommentRequest": {
                "required": [
                    "text"
                ],
                "properties": {
                    "text": {
                        "type": "string",
                        "example": "Updated comment text"
                    }
                },
                "type": "object"
            },
            "CommentResource": {
                "title": "Comment API Resource",
                "description": "Public API representation of comment",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/CommentModel"
                    },
                    {
                        "properties": {
                            "user": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    }
                                },
                                "type": "object",
                                "nullable": true
                            },
                            "replies": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "CommentModel": {
                "title": "Comment",
                "description": "Comment model with support for nested comments",
                "required": [
                    "user_id",
                    "commentable_type",
                    "commentable_id",
                    "text"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "commentable_type": {
                        "type": "string",
                        "example": "App\\Services\\News\\Models\\News"
                    },
                    "commentable_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "text": {
                        "type": "string",
                        "example": "This is a comment"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-15T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-15T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}